Dayframe (Fullstack)/requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-WTF==1.2.2
email-validator==2.3.0
WTForms==3.2.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
Werkzeug==3.1.3
python-dotenv==1.2.1
itsdangerous==2.2.0
tzdata==2025.2